Add tiledimage module to calculate samples of multi res images
It calculates which tiles are required to render a given region of
a cudl image, given the size of the image at a certain level.

We can fetch the size of the image at a certain level by fetching the thumbnail image (lvl 8), and using that as a preview, and as a basis
for calculating the tiles to be fetched.